requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pdfminer.six>=20221105
PyPDF2>=3.0.0
